feat: improve performance for pods and collabs
