+++ wondrous-app/components/Common/TaskViewModal/TaskViewModalWatcher.tsx
@@ -1,0 +1,28 @@
+import React, { Suspense } from 'react';
+import dynamic from 'next/dynamic';
+import { useRouter } from 'next/router';
+
+import { ViewType } from 'types/common';
+
+const TaskViewModal = dynamic(() => import('components/Common/TaskViewModal'), { suspense: true });
+
+const TaskViewModalWatcher = () => {
+  const router = useRouter();
+  const { query } = router;
+
+  const taskId = (query.task || query.taskProposal) as string | null;
+  const handleClose = () => {
+    delete query.task;
+    delete query.taskProposal;
+
+    router.push({ query }, undefined, { scroll: false });
+  };
+
+  return taskId ? (
+    <Suspense>
+      <TaskViewModal open handleClose={handleClose} taskId={taskId} back={query.view === ViewType.List} />
+    </Suspense>
+  ) : null;
+};
+
+export default TaskViewModalWatcher;
