+++ wondrous-app/components/Dashboard/boards/BoardSkeleton.tsx
@@ -10,15 +10,33 @@ const BoardSkeleton = () => {
   const styles = { bgcolor: palette.grey900 };
 
   return (
-    <Box mt="30px">
-      <Grid container wrap="nowrap" gap="20px" mt="10px" columns={4}>
+    <Box px="50px" mt="160px">
+      <Grid container wrap="nowrap" gap="20px" columns={4} my="20px">
         {Array.from(new Array(4)).map((index) => (
-          <Grid item sm={3} key={index}>
-            <Skeleton sx={styles} variant="rectangular" width="50%" height={30} />
+          <Skeleton key={index} sx={styles} variant="rectangular" width={80} height={45} />
+        ))}
+      </Grid>
+
+      <Skeleton sx={{ bgcolor: palette.grey75 }} variant="rectangular" width="100%" height={3} />
+
+      <Grid container wrap="nowrap" gap="20px" columns={4} my="20px">
+        <Grid item sm={6}>
+          <Skeleton sx={styles} variant="rectangular" width="100%" height={45} />
+        </Grid>
+        {Array.from(new Array(3)).map((index) => (
+          <Grid item sm={2} key={index}>
+            <Skeleton sx={styles} variant="rectangular" width="100%" height={45} />
           </Grid>
         ))}
       </Grid>
 
+      <Grid container wrap="nowrap" gap="20px" columns={4}>
+        {Array.from(new Array(4)).map((index) => (
+          <Grid key={index} item sm={3}>
+            <Skeleton sx={styles} variant="rectangular" width="100%" height={180} />
+          </Grid>
+        ))}
+      </Grid>
       <Grid container wrap="nowrap" gap="20px" mt="10px" columns={4}>
         {Array.from(new Array(4)).map((index) => (
           <Grid key={index} item sm={3}>
