+++ wondrous-app/components/ListView/Item.tsx
@@ -284,9 +284,12 @@ export default function ListViewItem({ task, entityType }) {
         preventLinkNavigation
         onNavigate={() => {
           if (!showPaymentModal) {
-            // location.push(viewUrl);
-            // windowOffset = window.scrollY;
-            // document.body.setAttribute('style', `position: fixed; top: -${windowOffset}px; left:0; right:0`);
+            const query = {
+              ...router.query,
+              [taskType]: task?.id
+            }
+
+            router.push({ query }, undefined, { scroll: false, shallow: true });
           }
         }}
       >
