+++ wondrous-app/components/Pod/analytics/index.tsx
@@ -79,7 +79,6 @@ function UserRow({ contributorTask }) {
     <ContributorDiv>
       <TaskViewModal
         open={openModal}
-        shouldFocusAfterRender={false}
         handleClose={() => {
           const style = document.body.getAttribute('style');
           const top = style.match(/(top: -)(.*?)(?=px)/);
