+++ wondrous-app/components/organization/analytics/index.tsx
@@ -164,7 +164,6 @@ function UserRow({ contributorTask }) {
     <ContributorDiv>
       <TaskViewModal
         open={openModal}
-        shouldFocusAfterRender={false}
         handleClose={() => {
           const style = document.body.getAttribute('style');
           const top = style.match(/(top: -)(.*?)(?=px)/);
